Update level.rst
when searching the docs for "evidence level" this page is low on the search result list, as the top level headers are just "level". Since these are _evidence_ levels, I think it would improve search performance and readability, at the cost of some structural redundancy (as this topic is nested inside the `evidence` topic).